+++ src/demo_nft_marketplace/handlers/on_cancel_swap.py
@@ -9,6 +9,6 @@ async def on_cancel_swap(
     ctx: HandlerContext,
     cancel_swap: TezosTzktTransaction[CancelSwapParameter, HenMinterStorage],
 ) -> None:
-    swap = await models.Swap.filter(id=int(cancel_swap.parameter.__root__)).get()
+    swap = await models.Swap.filter(id=int(cancel_swap.parameter.root)).get()
     swap.status = models.SwapStatus.CANCELED
     await swap.save()