+++ src/dipdup/config/evm.py
@@ -2,7 +2,7 @@ from typing import Literal
 
 from eth_utils.address import is_address
 from eth_utils.address import to_normalized_address
-from pydantic import validator
+from pydantic import field_validator
 from pydantic.dataclasses import dataclass
 
 from dipdup.config import ContractConfig
@@ -27,7 +27,8 @@ class EvmContractConfig(ContractConfig):
     abi: str | None = None
     typename: str | None = None
 
-    @validator('address', 'abi', allow_reuse=True)
+    @field_validator('address', 'abi')
+    @classmethod
     def _valid_address(cls, v: str | None) -> str | None:
         # NOTE: It's a `config export` call with environment variable substitution disabled
         if not v or '$' in v: