+++ src/dipdup/indexes/tezos_tzkt_operations/parser.py
@@ -4,6 +4,7 @@ from functools import lru_cache
 from itertools import groupby
 from types import UnionType
 from typing import Any
+from typing import Optional
 from typing import TypeVar
 from typing import Union
 from typing import cast
@@ -11,7 +12,6 @@ from typing import get_args
 from typing import get_origin
 
 from pydantic import BaseModel
-from pydantic import Extra
 
 from dipdup.exceptions import InvalidDataError
 from dipdup.models.tezos_tzkt import TezosTzktOperationData
@@ -25,12 +25,12 @@ IntrospectionError = (KeyError, IndexError, AttributeError)
 
 def extract_root_outer_type(storage_type: type[BaseModel]) -> type[BaseModel]:
     """Extract Pydantic __root__ type"""
-    root_field = storage_type.__fields__['__root__']
-    if root_field.allow_none:
+    root_field = storage_type.model_fields['root']
+    if not root_field.is_required():
         # NOTE: Optional is a magic _SpecialForm
-        return cast(type[BaseModel], root_field.type_ | None)
+        return cast(type[BaseModel], Optional[root_field.annotation])  # noqa: UP007
 
-    return root_field.outer_type_  # type: ignore[no-any-return]
+    return root_field.annotation  # type: ignore[return-value]
 
 
 def is_array_type(storage_type: type[Any]) -> bool:
@@ -39,7 +39,7 @@ def is_array_type(storage_type: type[Any]) -> bool:
     if get_origin(storage_type) == list:
         return True
 
-    # NOTE: Pydantic model with __root__ field subclassing List
+    # NOTE: Pydantic model with root field subclassing List
     with suppress(*IntrospectionError):
         root_type = extract_root_outer_type(storage_type)
         return is_array_type(root_type)
@@ -54,7 +54,7 @@ def get_list_elt_type(list_type: type[Any]) -> type[Any]:
     if get_origin(list_type) == list:
         return get_args(list_type)[0]  # type: ignore[no-any-return]
 
-    # NOTE: Pydantic model with __root__ field subclassing List
+    # NOTE: Pydantic model with root field subclassing List
     root_type = extract_root_outer_type(list_type)
     return get_list_elt_type(root_type)
 
@@ -65,7 +65,7 @@ def get_dict_value_type(dict_type: type[Any], key: str | None = None) -> type[An
     if get_origin(dict_type) == dict:
         return get_args(dict_type)[1]  # type: ignore[no-any-return]
 
-    # NOTE: Pydantic model with __root__ field subclassing Dict
+    # NOTE: Pydantic model with root field subclassing Dict
     with suppress(*IntrospectionError):
         root_type = extract_root_outer_type(dict_type)
         return get_dict_value_type(root_type, key)
@@ -74,13 +74,9 @@ def get_dict_value_type(dict_type: type[Any], key: str | None = None) -> type[An
         raise KeyError('Field name or alias is required for object introspection')
 
     # NOTE: Pydantic model, find corresponding field and return it's type
-    fields = dict_type.__fields__
-    for field in fields.values():
-        if key in (field.name, field.alias):
-            # NOTE: Pydantic does not preserve outer_type_ for Optional
-            if field.allow_none:
-                return cast(type[Any], field.type_ | None)
-            return field.outer_type_  # type: ignore[no-any-return]
+    for name, field in dict_type.model_fields.items():
+        if key in (name, field.alias):
+            return field.annotation  # type: ignore[no-any-return]
 
     # NOTE: Either we try the wrong Union path or model was modifier by user
     raise KeyError(f'Field `{key}` not found in {dict_type}')
@@ -159,16 +155,10 @@ def _process_storage(
 
     # NOTE: Dict, process recursively
     elif isinstance(storage, dict):
-        # NOTE: Ignore missing fields along with extra ones
-        ignore = getattr(getattr(storage_type, 'Config', None), 'extra', None) == Extra.ignore
-
         for key, value in storage.items():
-            try:
-                value_type = get_dict_value_type(storage_type, key)
-                storage[key] = _process_storage(value, value_type, bigmap_diffs)
-            except IntrospectionError:
-                if not ignore:
-                    raise
+            value_type = get_dict_value_type(storage_type, key)
+            storage[key] = _process_storage(value, value_type, bigmap_diffs)
+
     # NOTE: Leave others untouched
     else:
         pass