+++ tests/types/qwer/storage.py
@@ -3,23 +3,24 @@
 
 from __future__ import annotations
 
+from typing import Any
+
 from pydantic import BaseModel
-from pydantic import Extra
+from pydantic import ConfigDict
+from pydantic import RootModel
 
 
 class QwerStorageItem(BaseModel):
-    class Config:
-        extra = Extra.forbid
+    model_config = ConfigDict(extra='forbid')
 
     L: str
 
 
 class QwerStorageItem1(BaseModel):
-    class Config:
-        extra = Extra.forbid
+    model_config = ConfigDict(extra='forbid')
 
     R: dict[str, str]
 
 
-class QwerStorage(BaseModel):
-    __root__: list[list[QwerStorageItem | QwerStorageItem1]]
+class QwerStorage(RootModel[Any]):
+    root: list[list[QwerStorageItem | QwerStorageItem1]]