+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Changed
+
+- dipdup: Python 3.12 is now required to run DipDup.
+- dipdup: Pydantic v2 migration.
+
 ## [7.5.3] - 2024-03-28
 
 ### Fixed